+++ docs/source/_extensions/list_drivers.py
@@ -1,6 +1,6 @@
 """
 YAML GDAL/OGR driver list extension module for sphinx documentation
-John Truckenbrodt 2019
+John Truckenbrodt 2019, 2026
 
 With this, spatialist's yaml files listing known file extensions can be read and listed in the sphinx documentation.
 The argument is either 'vector' or 'raster' listing the content of files 'drivers_vector.yml' and 'drivers_raster.yml'
@@ -16,7 +16,7 @@ the following file extensions are auto-detected:
 import yaml
 from docutils import nodes
 from docutils.parsers.rst import Directive
-from pkg_resources import resource_filename
+from importlib.resources import files
 
 
 class ListDrivers(Directive):
@@ -24,8 +24,8 @@ class ListDrivers(Directive):
     
     def run(self):
         base = 'drivers_{}.yml'.format(self.arguments[0])
-        filename = resource_filename('spatialist', base)
-        drivers = yaml.safe_load(open(filename))
+        with files('spatialist').joinpath(base).open('r') as f:
+            drivers = yaml.safe_load(f)
         
         lst = nodes.bullet_list()
         for extension, name in drivers.items():