+++ docs/source/conf.py
@@ -15,7 +15,7 @@
 import sys
 import os
 import datetime
-from pkg_resources import get_distribution
+from importlib.metadata import version as read_version
 
 project = 'spatialist'
 authors = 'John Truckenbrodt, Ismail Baris, Felix Cremer, Marco Wolsza, Tyler Anderson'
@@ -35,10 +35,10 @@ sys.path.insert(0, os.path.join(os.path.abspath('.'), '_extensions'))
 # release = get_distribution('spatialist').version
 # The short X.Y version.
 # version = '.'.join(release.split('.')[:2])
-version = get_distribution(project).version
+version = read_version(project)
 
 # -- General configuration ------------------------------------------------
-autodoc_mock_imports = ['osgeo', 'matplotlib', 'mpl_toolkits', 'IPython', 'ipywidgets', 'sqlite3']
+autodoc_mock_imports = ['matplotlib', 'mpl_toolkits', 'IPython', 'ipywidgets', 'sqlite3']
 
 # If your documentation needs a minimal Sphinx version, state it here.
 needs_sphinx = '1.6'
@@ -53,6 +53,7 @@ extensions = [
     'sphinx.ext.napoleon',
     'sphinx.ext.intersphinx',
     'sphinx.ext.viewcode',
+    'sphinx_autodoc_typehints',
     'epsg_links',
     'list_drivers'
 ]